forced cast

+++ test/BSONElement_test.cpp
@@ -70,7 +70,7 @@ unittest(BSONElement_Constructor_t)
 unittest(BSONElement_double_t)
 {
     BSONElement be;
-    be.Key("pi").Value(3.1416);
+    be.Key("pi").Value((float)3.1416);
 
     /*
     *   \x01                              -> 0x01 = double
@@ -81,7 +81,7 @@ unittest(BSONElement_double_t)
     assertEqual(12, be.len());
     assertEqual(BSON_TYPE_NUMBER, be.getType());
     assertEqual("pi", be.getKey());
-    assertEqual(3.1416, be.getDouble());
+    assertEqual((float)3.1416, be.getDouble());
     assertFalse(be.isString());
     assertFalse(be.isInt());
     assertFalse(be.isBool());
